refactor to match lint rule (method length reduced)
