+++ packages/plugins/Gfi/src/store/actions/index.ts
@@ -5,19 +5,13 @@ import Overlay from 'ol/Overlay'
 import { GeoJSON } from 'ol/format'
 import { Feature } from 'ol'
 import { Feature as GeoJsonFeature, GeoJsonProperties } from 'geojson'
-import { rawLayerList } from '@masterportal/masterportalapi'
 import { PolarActionTree } from '@polar/lib-custom-types'
 import getCluster from '@polar/lib-get-cluster'
 import { getTooltip, Tooltip } from '@polar/lib-tooltip'
-import {
-  getFeatureDisplayLayer,
-  clear,
-  addFeature,
-} from '../../utils/displayFeatureLayer'
-import { requestGfi } from '../../utils/requestGfi'
+import { getFeatureDisplayLayer, clear } from '../../utils/displayFeatureLayer'
 import { GfiGetters, GfiState } from '../../types'
-import sortFeatures from '../../utils/sortFeatures'
 import { getOriginalFeature } from '../../utils/getOriginalFeature'
+import { debouncedGfiRequest } from './debouncedGfiRequest'
 
 // OK for module action set creation
 // eslint-disable-next-line max-lines-per-function
@@ -225,117 +219,7 @@ export const makeActions = () => {
       // call further stepped in a debounced fashion to avoid a mess
       return await dispatch('debouncedGfiRequest', coordinate)
     },
-    /**
-     * Code from `getFeatureInfo`, pulled to avoid overly requesting feature
-     * information. Since sources in Pins plugin update right after each other
-     * (and such effects are to be expected across the system), we're debouncing
-     * this *after* resetting the module state, as something is bound to happen.
-     */
-    debouncedGfiRequest: debounce(
-      // TODO: Types are not properly displayed here as it is wrapped through debounce
-      async (
-        {
-          commit,
-          rootGetters: { map, configuration },
-          getters: { layerKeys, geometryLayerKeys, afterLoadFunction },
-        },
-        coordinate: [number, number]
-      ): Promise<void> => {
-        // fetch new feature information for all configured layers
-        const promisedFeatures: Promise<GeoJsonFeature[]>[] = layerKeys.map(
-          (key) => {
-            const layer = map
-              .getLayers()
-              .getArray()
-              .find((layer) => layer.getProperties().id === key)
-
-            if (!layer) {
-              console.error(
-                `@polar/plugin-gfi: No layer with id "${key}" found during run-time. GFI skipped.`
-              )
-              return [] as GeoJsonFeature[]
-            }
-
-            const layerConfiguration = configuration.gfi.layers[key] || {}
-            const layerSpecification = rawLayerList.getLayerWhere({ id: key })
-            const mainLayerConfiguration = configuration.layers.find(
-              (element) => element.id === key
-            )
-            const layerGfiMode =
-              mainLayerConfiguration.gfiMode ||
-              configuration.gfi.mode ||
-              'bboxDot'
-
-            return requestGfi({
-              map,
-              layer,
-              coordinate,
-              layerConfiguration,
-              layerSpecification,
-              mode: layerGfiMode,
-            })
-          }
-        )
-
-        const errorSymbol = (err) => Symbol(err)
-        const features = (await Promise.allSettled(promisedFeatures)).map(
-          (result) =>
-            result.status === 'fulfilled'
-              ? result.value
-              : errorSymbol(result.reason.message)
-        )
-
-        const generalMaxFeatures: number =
-          configuration.gfi.maxFeatures || Number.POSITIVE_INFINITY
-
-        const srsName: string = map.getView().getProjection().getCode()
-        // map features back to their layer keys
-        let featuresByLayerId: Record<string, GeoJsonFeature[] | symbol> =
-          layerKeys.reduce(
-            (accumulator, key, index) => ({
-              ...accumulator,
-              [key]: Array.isArray(features[index])
-                ? (features[index] as []).slice(
-                    0,
-                    configuration.gfi.layers[key].maxFeatures ||
-                      generalMaxFeatures
-                  )
-                : features[index],
-            }),
-            {}
-          )
-        featuresByLayerId = Object.entries(featuresByLayerId).reduce(
-          (accumulator, [layerKey, layerValues]) => ({
-            ...accumulator,
-            [layerKey]:
-              Array.isArray(layerValues) && layerValues.length >= 2
-                ? layerValues.sort((a, b) => sortFeatures(a, b, srsName))
-                : layerValues,
-          }),
-          {}
-        )
-
-        // store features in state, if configured via client specific function
-        if (typeof afterLoadFunction === 'function') {
-          featuresByLayerId = await afterLoadFunction(
-            featuresByLayerId,
-            srsName
-          )
-        }
-        commit('setFeatureInformation', featuresByLayerId)
-
-        // render feature geometries to help layer
-        geometryLayerKeys
-          .filter((key) => Array.isArray(featuresByLayerId[key]))
-          .forEach((key) =>
-            // @ts-expect-error | Might be fixed through having all the types in the action. Otherwise: It works properly, as all the symbols are filtered before calling forEach
-            featuresByLayerId[key].forEach((feature) =>
-              addFeature(feature, featureDisplayLayer)
-            )
-          )
-      },
-      50
-    ),
+    debouncedGfiRequest: debouncedGfiRequest(featureDisplayLayer),
     setCoreSelection(
       { commit, dispatch, rootGetters },
       {
